Add files via upload
Requirements:

You will submit a single ZIP archive containing the following files:

A PDF file containing a print-quality map of digitized buildings in the area covered by the Sanborn map of Hollywood.
A folder containing the shapefile of digitized buildings that you created for this assignment (note: all components of the shapefile, including .shp, .shx, and .dbf, must be included in the archive).
The following elements are required in your shapefile:

Digitized outlines of all buildings and/or establishments contained in the Sanborn map, with the exception of open spaces;
A field titled CLASS corresponding to building function/type.  Every feature on the map must contain an entry for this field.  The following one-character values should be used:  R for residential structures, C for commercial structures, I for industrial structures, P for public structures, L for landmarks that do not otherwise correspond to the aforementioned categories, and U for structures whose function cannot be easily read from the map.
A field titled NOTES where you will enter any text related to the structure marked on the Sanborn map, with the exception of measurements, construction notes, and one-letter structure codes (e.g. S and D).  Multiple annotations on a feature should be separated by commas.  When digitizing the Mark Twain Hotel's structure, for example, you would enter "Mark Twain Hotel, 70 Rms".
The following elements are required in your map:

Digitized outlines of all buildings and/or establishments contained in the Sanborn map, with fill color determined on the basis of values in the CLASS field (e.g. features with "R" in the class field would be filled with one color, and features with "C" in the field would be filled with a different color);
Outlines of street blocks and labels indicating street names drawn upon your map using manual annotations;
Standard required map elements, including a legend indicating the different categories from the CLASS field.
Keep in mind the following notes and guidelines:

Structures labeled with "S" on the Sanborn map are generally stores or structures being used for a commercial purpose, while structures labeled with "D" are generally dwellings or structures being used for residential purposes.
Dotted lines indicate internal divisions within structures.  These internal divisions correspond to separate structural entities, and should be digitized as distinct features.
Solid lines adjacent to streets do not necessarily indicate presence of a building, but rather indicate the bounds of a single street block (see the numerous open areas labeled "AUTO PARK", for example).  Open spaces and auto parks (parking lots) should not be digitized.