Add files via upload
Requirements:

Your task is to use overlay analysis to create a series of three maps related to fire hazard in the city of Riverside, California.  Each map will require you to use a unique combination of geoprocessing and overlay operations.  You will be provided with a list of the key operations that you will need to use to create each map, but it is your responsibility to define the workflow that you will use to create each one of the maps.  All maps must reflect principles of high-quality cartographic design and should contain (1) common required map elements, (2) your name, (3) the outline of the city of Riverside, and (3) an appropriate basemap.

Several notes about the source data and the terminology used in the assignment specifications:

An area is said to be classified with a certain level of fire hazard if the fire hazard zone's description is "Moderate", "High", or "Very High".  Other areas may be included in the source data but are not considered to represent classified fire hazard zones.
Maps of fire hazard produced by CAL FIRE (Links to an external site.) represent "Moderate" fire hazard zones with a yellow fill color, "High" zones with an orange fill color, and "Very High" zones with a red fill color.  It is recommended that your maps use this same color scheme when and where appropriate.
You can assume that residential parcels in the city of Riverside are any parcels whose zone description contains the term "Residential".
MAP 1:  Overview of fire hazard in Riverside, California (clip, spatial selection)

This map should display fire hazard zones within Riverside only (fire hazard outside the city should not be visible), and the zones should be shaded on the basis of their zone class (Moderate, High, or Very High).  The map should also display named point landmarks (i.e., landmarks without names present can be excluded) within the city that are contained within the fire hazard zones, and these at-risk landmarks should be labeled by name, when possible and appropriate.
MAP 2:  Choropleth map of Riverside neighborhoods, categorized by percentage of area within classified fire hazard zones (dissolve, clip or intersect, attribute query/filtering, attribute calculation, attribute join)

This map should display the extent of neighborhoods within the City of Riverside, and each neighborhood should be shaded on the basis of the percentage of its coverage area lying within fire hazard zones (Moderate, High, and Very High combined).  You will need to use geoprocessing operations on the neighborhood and fire hazard layers in order to obtain these area measurements.  The choropleth map that you create should use the an appropriate classification scheme with at least four data categories. Which neighborhood has the third highest fire hazard percentage? Add your answer (neighborhood name + rounded percentage) to your map as a caption.
MAP 3:  Residential parcels categorized by fire hazard (clip or intersect, spatial join)

The purpose of this map is to present the residential parcels within the city of Riverside that are classified as being within a moderate, high, or very high fire hazard zone.  On one PDF page (you may choose the page size -- large-format is fine if it works better), the map should contain (1) a citywide map showing those parcels intersecting a moderate, high, or very high fire hazard zone, and (2) at least two neighborhood-level inset maps, focused on several of the most fire-vulnerable areas of the city, visually highlighting those parcels within classified fire hazard areas.  Numbers and percentages of residential parcels belonging to each of the fire hazard areas should be displayed for the full city and for each of the selected neighborhoods, either in tabular form (on the map page) or using captions placed on the map.
Note that this week's technical videos demonstrate all of the new techniques that you will need to complete this week's assignment -- but the exact procedures required to produce these maps are not demonstrated.  You are responsible for using this week's videos and your experience with GIS to complete this assignment!